+++ spec/nanook/block_spec.rb
@@ -409,7 +409,7 @@ RSpec.describe Nanook::Block do
     expect(Nanook.new.block(block).successors(limit: 1000)).to be_empty
   end
 
-  it 'should request is_valid_work? correctly when valid' do
+  it 'should request valid_work? correctly when valid' do
     work = '2bf29ef00786a6bc'
 
     stub_request(:post, uri).with(
@@ -421,10 +421,10 @@ RSpec.describe Nanook::Block do
       headers: {}
     )
 
-    expect(Nanook.new.block(block).is_valid_work?(work)).to be true
+    expect(Nanook.new.block(block).valid_work?(work)).to be true
   end
 
-  it 'should request is_valid_work? correctly when not valid' do
+  it 'should request valid_work? correctly when not valid' do
     work = '2bf29ef00786a6bc'
 
     stub_request(:post, uri).with(
@@ -436,6 +436,6 @@ RSpec.describe Nanook::Block do
       headers: {}
     )
 
-    expect(Nanook.new.block(block).is_valid_work?(work)).to be false
+    expect(Nanook.new.block(block).valid_work?(work)).to be false
   end
 end